dolar_project/requirements.txt
asgiref==3.7.2
Django==4.2.6
sqlparse==0.4.4
typing_extensions==4.8.0
psycopg2-binary
pandas
requests
django-cors-headers
prophet==1.1.5
matplotlib